+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.0]
+### Changed
+- Wrap clock js in class
+- Add PHP example index
+
 ## [0.1.1]
 ### Changed
 - Updated build